v2/instagram_song_identifier/requirements.txt
instagrapi
moviepy
requests
python-dotenv